Allow GUI conversion to signed R/RG uncompressed texture formats
Allow conversion to signed R and RG uncompressed texture formats
in the GUI program. The resulting uncompressed texture can be saved
as a KTX or DDS file. Such texture files can be loaded as as source
for a texture compression format that requires 16-bit input, such as
signed_rgtc1/2. Update README.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>