Reorganize speed/quality level settings
Reorganize quality level settings by adding --level option with a
level in the range 0-50. The presets --ultra, --fast, --medium and
--slow correspond to a particular level.

Implement a second pass for compression in which a small population
size is used (8) and run for a large number of generations. The
second pass uses a relatively small amount of time due to the small
population size. The second pass is intended as a final fine-tuning
step.

Reorder options in texgenpack command-line utility.

Update README.

Bump version to 0.8.1.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>