Provide perceptive quality strategy
Provide perceptive quality strategy whereby not just the RGB
differences of the pixels within a block are taken into account,
but also the differences between the RGB differences between
adjacent pixels, as well as adjacent pixels in the block immediately
above and to the left of the evaluated block.

For example, consider the case of two adjacent pixels having a component
value of 128 in the original texture. In one compressed texture
evaluated by the genetic algorithm, the component values are 127 and
129, and in another they are both 127. In terms of pure RGB difference
(RMSE), the two compressed solutions wil have identical fitness, but
with a perceptive strategy, the solution where both pixels have a
component value of 127 will be preferred.

Although this strategy can produce higher RMSE (higher divergence
from the original texture in terms of pure RGB differences), the
visual (perceptive) quality may be better. The strategy also
addresses discontinuities between pixels in adjacent blocks, because
the fitness evaluation incorporates pixel value differences for the
top row of a block with the bottom row of the block above, and for the
left column of a block with the rightmost column block to the left.

The perceptive quality strategy is currently only implemented for RGB8
texture formats (no alpha). For other formats, the standard (non-
perceptive) quality strategy will be followed.

The --ultra class of quality level settings (0 to 7) cannot use the
perceptive strategy because multiple different blocks are compressed
in parallel. However, for higher quality level settings (8 to 50,
corresponding to fast, medium and slow classes), the perceptive
strategy is now the default. The --non-perceptive option has been
added to allow selection of the standard non-perceptive strategy.

This commit also reworks the second pass for compression levels >= 8,
so that it is now integrated within the first pass. Instead of running
the second pass on a single genetic algorithm population (island) for
each block, the second pass now uses four seperate islands (the best
result is selected), and the number of generations for the second pass
has been reduced. The --generations-second-pass and
--islands-second-pass options ahve been added to allow specific control
of second pass parameters.

The --maxthreads option is handled more gracefully by more
fully utilizing the allowed number of threads, instead of defaulting
to a single thread when the intended concurrency is too high.

Bump version to v0.9.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>