Improve compression efficiency and quality
Improve compression efficiency/quality for textures with alpha by
predetermining whether the block alpha is opaque or non-opaque
and subsequently allowing/disallowing specific modes of the compression
format. Applies to ETC2_PUNCHTHROUGH and BPTC textures.

Keep track of modes used with the BPTC format and add the capability to
allow/disallow specific modes in the decoding function.

Make the modal genetic algorithm strategy (where each of the islands is
tied to specific modes) the default also for for ETC_EAC and
ETC2_PUNCHTHROUGH.

For ETC1, ETC2, BPTC and BPTC_FLOAT, preserve the mode of the compressed
block found in the first pass during the second pass.

Fix problems when the number of components of the source image and
compressed texture format differ (extra alpha component).

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>