Improve BPTC compression with two-color detection
For the BPTC compression format, use color information when only two
colors are used inside a block to initialize the compressed block colors
during the seeding function, depending on the BPTC mode. Implemented
for BPTC modes 5 and 6 (with alpha) and mode 3 (opaque).

The same technique could be used to improve compressions for other
formats.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>