Fix loading of uncompressed textures with odd number of row bytes
For certain uncompressed texture formats, such as R8, RG8, R16 and
R_HALF_FLOAT, the KTX texture loading function was returning an
error message if the row size is not a multiple of four.

Apart from non-power-of-two textures, this bug also affected loading
of power-of-two textures in the mentioned formats when they include
mipmap levels with a row size smaller than four bytes.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>